razer-dashboard/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
docker==7.0.0
psutil==5.9.8
requests==2.31.0
pyyaml==6.0.1
pydantic==2.5.3
watchdog==3.0.0
python-multipart==0.0.6
